It's done.
Once upon a time, there was an university library.
The library was sad, because it didn't have an online system.
But then two programming knights came to build such a system.
Over months, in countless hours, they handcrafted such a system,
Out of a poorly written codebase of a CODE orientation semester project.
Today the first component was finished and the library was happy.
From now on, she could show all people her real worth...